Move wasm-only code into wasm build flags
